+++ src/app/main-page/request/pettycash/pettycashrequest/page.tsx
@@ -2,13 +2,14 @@
 
 import VoucherBlue from "./components/VoucherBlue/VoucherBlue";
 import VoucherPink from "./components/VoucherPink/VoucherPink";
+import { PettyCashProvider } from "./context/PettyCashProvider";
 
 const PettyCashRequest = () => {
   return (
-    <>
+    <PettyCashProvider>
       <VoucherPink />
       <VoucherBlue />
-    </>
+    </PettyCashProvider>
   );
 };